% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{contour.plot.helper}
\alias{contour.plot.helper}
\title{Filled contour plot helper}
\usage{
\method{contour}{plot.helper}(x, ngrid = 500, counter_levels = 8, ...)
}
\description{
Filled contour plot helper
}
\keyword{internal}
