% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_frontier.R
\name{estimate_frontier}
\alias{estimate_frontier}
\title{Estimate the production frontier f(x)}
\usage{
estimate_frontier(g_hat, p_hat)
}
\arguments{
\item{g_hat}{Numeric vector of estimated g(x).}

\item{p_hat}{Numeric scalar estimate of parameter p.
$p$ has to be greater than zero, as cited in the article}
}
\value{
Numeric vector of estimated frontier values f(x).
}
\description{
Estimate the production frontier f(x)
}
