% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsuoka_density.R
\name{matsuoka.density}
\alias{matsuoka.density}
\title{Implements Matsuoka density function}
\usage{
matsuoka.density(x, p)
}
\arguments{
\item{x}{numeric vector of values \eqn{x \contains (0,1)}}

\item{p}{positive parameter.}
}
\value{
numeric vector of density values.
}
\description{
Implements Matsuoka density function
}
