% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatsuokaFrontier.R
\name{MatsuokaFrontier}
\alias{MatsuokaFrontier}
\title{MatsuokaFrontier Class}
\usage{
MatsuokaFrontier(x, y)
}
\arguments{
\item{x}{A numeric data frame or matrix representing the independent variables.
If \code{x} is a data frame, the columns should be numeric.}

\item{y}{A numeric vector representing the dependent variable.}
}
\value{
An object of class \code{MatsuokaFrontier} containing:
\describe{
\item{x}{The input data frame of independent variables.}
\item{y}{The dependent variable vector.}
\item{estimated_p}{The estimated parameter \code{p}.}
\item{estimated_g}{The estimated function \code{g(x)}.}
}
}
\description{
This class estimates the production frontier based on the provided inputs.
It can handle both univariate and multivariate input data (\code{x}).
}
