% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.matsuoka3step}
\alias{plot.matsuoka3step}
\title{Plot method for matsuoka3step objects}
\usage{
\method{plot}{matsuoka3step}(obj, ngrid = 500, counter_levels = 8, ...)
}
\arguments{
\item{obj}{An object of class \code{matsuoka3step}.}

\item{ngrid}{Number of grid points per axis.}

\item{counter_levels}{levels of contour from the plot.}

\item{...}{Extra graphical arguments passed to filled.contour.}
}
\description{
Plot method for matsuoka3step objects
}
