% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsuoka.R
\name{rmatsuoka}
\alias{rmatsuoka}
\title{Random generation from the Matsuoka distribution}
\usage{
rmatsuoka(n, p)
}
\arguments{
\item{n}{number of observations}

\item{p}{positive shape parameter}
}
\value{
numeric vector of length n with random draws
}
\description{
Random generation from the Matsuoka distribution
}
