% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.matsuoka3step}
\alias{summary.matsuoka3step}
\title{Summary for matsuoka3step}
\usage{
\method{summary}{matsuoka3step}(object, digits = 4, ...)
}
\arguments{
\item{object}{A 'matsuoka3step' object.}

\item{digits}{Number of digits to print.}

\item{...}{unused}
}
\description{
Summary for matsuoka3step
}
