% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate.X}
\alias{validate.X}
\title{Validate that X is a data.frame}
\usage{
validate.X(X)
}
\arguments{
\item{X}{An input object}
}
\value{
Stops with error if invalid
}
\description{
Validate that X is a data.frame
}
\keyword{internal}
