% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate.y}
\alias{validate.y}
\title{Validate that y is numeric and strictly positive}
\usage{
validate.y(y)
}
\arguments{
\item{y}{Numeric vector}
}
\value{
Stops with error if invalid
}
\description{
Validate that y is numeric and strictly positive
}
\keyword{internal}
