% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.R
\name{cvm_matsuoka_test}
\alias{cvm_matsuoka_test}
\title{Cramér-von Mises criterion test for the Matsuoka distribution}
\usage{
cvm_matsuoka_test(
  x,
  p,
  estimated = TRUE,
  nullname = "Matsuoka Distribution",
  ...
)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{p}{Numeric scalar, the Matsuoka parameter.}

\item{estimated}{Logical, whether the parameter \code{p} is estimated from the data, default is set to TRUE.}

\item{nullname}{Character, name of the null distribution for reporting.}

\item{...}{Additional arguments passed to \code{\link[goftest]{cvm.test}}.}
}
\value{
An object of class \code{htest}, as returned by \code{cvm.test()}.
}
\description{
Performs the Cramér-von Mises goodness-of-fit test to check whether
a numeric vector \code{x} follows a Matsuoka distribution.
}
\examples{
# In practice, x_sample would be obtained as y / f_hat after using matsuoka3step.
# For reproducibility in examples, we simulate Matsuoka data using `rmv()` function:
x_sample <- rmv(100, p = 0.5)

# Perform Cramér–von Mises test
cvm_matsuoka_test(x_sample, p = 0.5)

# Pass additional arguments to cvm.test()
cvm_matsuoka_test(x_sample, p = 0.5, estimated = FALSE)

}
\seealso{
\code{\link[goftest]{cvm.test}} for the underlying test implementation.
\code{\link{visual_cvm_matsuoka}} for graphical assessment of the test.
}
