% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_p.R
\name{estimate.p}
\alias{estimate.p}
\title{Estimate parameter p from residuals}
\usage{
estimate.p(residuals)
}
\arguments{
\item{residuals}{Numeric vector of residuals (z - g_hat).}
}
\value{
Numeric estimate of p.
}
\description{
Estimates the shape parameter \( p \) for Matsuoka's distribution
using the method of moments.
}
