% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Run the Three-Step Frontier Estimation}
\usage{
estimate(X, y, g_strategy = "ll", ...)
}
\arguments{
\item{X}{A data.frame of input variables}

\item{y}{A numeric output vector}

\item{g_strategy}{Method for estimating g(x). Options: "ll"
TODO: Options: "spline", or "scar"}

\item{...}{Additional arguments passed to specific \code{g} estimators.}
}
\value{
A list with elements: g_hat, residuals, p_hat, f_hat.
}
\description{
This function estimates the production frontier using the three-step method:
(1) Any nonparametric regression to estimate g(x),
(2) estimation of the parameter p via method of moments,
(3) construction of f(x) via plug-in.
}
