% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{f.hat.x.y.plot}
\alias{f.hat.x.y.plot}
\title{Plot each xy values with productive estimated function}
\usage{
f.hat.x.y.plot(res, ...)
}
\arguments{
\item{res}{A fitted object containing at least:
\describe{
\item{x}{A data frame of input variables.}
\item{y}{A numeric vector of observed outputs.}
\item{f_hat}{A numeric vector of fitted/estimated outputs.}
}}
}
\value{
Invisibly returns \code{NULL}. The function is called for its side
effect of producing plots.
}
\description{
Plot each xy values with productive estimated function
}
\details{
Axis labels can be set with \code{xlab} and \code{ylab}.
Defaults are \code{"x"} and \code{"y"} if not provided.
}
\keyword{Plot}
\keyword{This}
\keyword{\code{f_hat}.}
\keyword{\code{x},}
\keyword{\code{y}}
\keyword{against}
\keyword{along}
\keyword{corresponding}
\keyword{each}
\keyword{estimated}
\keyword{frontier}
\keyword{function}
\keyword{functions}
\keyword{in}
\keyword{input}
\keyword{internal}
\keyword{observed}
\keyword{output}
\keyword{plots}
\keyword{production}
\keyword{the}
\keyword{variable}
\keyword{with}
