% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsuoka.R
\name{F.mv.i}
\alias{F.mv.i}
\title{Quantile function of the Matsuoka distribution}
\usage{
F.mv.i(q, p, igamma_inv_fun = zipfR::Igamma.inv)
}
\arguments{
\item{q}{Numeric vector of probabilities (values in \link{0,1}).}

\item{p}{Positive numeric scalar, distribution parameter.}

\item{igamma_inv_fun}{Optional function to compute the inverse incomplete gamma.
Defaults to \code{Igamma.inv} from the zipfR package.}
}
\value{
Numeric vector of quantiles corresponding to \code{q}.
}
\description{
Computes the inverse CDF (quantile function) of the Matsuoka distribution
with parameter \code{p}.
}
