% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsuoka3step.R
\name{matsuoka3step}
\alias{matsuoka3step}
\title{matsuoka3step Class Constructor}
\usage{
matsuoka3step(x, y, g = "spline", ...)
}
\arguments{
\item{x}{A data.frame of input variables.}

\item{y}{A numeric vector of outputs (must be > 0).}

\item{g}{A strategy for estimating g(x) ("spline" or custom function).}

\item{...}{Extra args passed to g(x) estimator.}
}
\value{
An object of class \code{matsuoka3step}.
}
\description{
matsuoka3step Class Constructor
}
