% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.matsuoka3step}
\alias{summary.matsuoka3step}
\title{Summary for matsuoka3step}
\usage{
\method{summary}{matsuoka3step}(x, digits = 4, tol = 1e-08, ...)
}
\arguments{
\item{x}{A 'matsuoka3step' object.}

\item{digits}{Number of digits to print.}

\item{tol}{Tolerance to treat values as efficient (score >= 1 - tol).}

\item{...}{unused}
}
\description{
Summary for matsuoka3step
}
